Moved all modules to be dev dependencies. Any app/library using this node module is not going to use any of those node modules.
TL;DR - This not going to be used in a node environment. Since this is primarily for browser environment app/library trying to use this
doesn't require any of the node modules when doing `npm i cask-datavoyager`.

All that is being used is built by webpack into one single file to be imported. Anything other than
build folder is not necessary to any module using data voyager

This is primarily due to the frustration of node-sass. We had a postinstall step of `npm rebuild node-sass` which is completely
unecessary. Since `font-awesome-sass-loader` was declared as dev dependency `npm i cask-datavoyager` was bringing down the whole
node_modules folder which brought in node-sass. Because of node-sass the webpack build was hanging for weird reasons.

To avoid all these I have moved all modules to development dependencies making lib-adapter-voyager.js the only file of importance.